refactor(doctor): one shared workspace dir-rename chokepoint; admit .agents + workspace to capabilities
Consolidate the three per-lane directory-rename adapters
(workspaceQuarantineRename, quarantineEmptyDir, workspaceRenameDir) onto a
single workspaceDirRename with an under-lock verify hook; add .agents to
canonicalWriteScopes and workspace to canonicalSubsystems so the workspace
fixers pass the production envelope (drift-lane tests now run against
production capabilities verbatim).
